+++ CHANGELOG.md
@@ -1,4 +1,9 @@
+#### September 2020
+- PR [#143](https://github.com/uma-pi1/kge/pull/143): Allow processing of large batches in smaller subbatches to save memory
+
 #### August 2020
+- PR [#140](https://github.com/uma-pi1/kge/pull/140): Calculate penalty for entities only once, if subject-embedder == object-embedder
+- PR [#138](https://github.com/uma-pi1/kge/pull/138): Revision of hooks, fix of embedding normalization
 - PR [#135](https://github.com/uma-pi1/kge/pull/135): Revised sampling API, faster negative sampling with shared samples
 
 #### June 2020